requirements.txt
gradio==4.44.0
deep-translator==1.11.4
pyyaml==6.0.1
requests==2.31.0
langdetect==1.0.9
fastapi==0.115.6
audioop-lts==0.2.2; python_version >= "3.13"
